Move `binding.js` to repository root (#129)
