+++ binding.js
@@ -1,0 +1,1 @@
+module.exports = require('node-gyp-build')(__dirname)
