--- lib/binding.js
@@ -1,4 +1,0 @@
-const path = require('path')
-const load = require('node-gyp-build')
-
-module.exports = load(path.join(__dirname, '..'))
