+++ lib/network-interfaces.js
@@ -1,6 +1,6 @@
 const events = require('events')
 const b4a = require('b4a')
-const binding = require('./binding')
+const binding = require('../binding')
 
 module.exports = class NetworkInterfaces extends events.EventEmitter {
   constructor () {
