+++ lib/socket.js
@@ -1,6 +1,6 @@
 const events = require('events')
 const b4a = require('b4a')
-const binding = require('./binding')
+const binding = require('../binding')
 const ip = require('./ip')
 
 module.exports = class UDXSocket extends events.EventEmitter {
