+++ lib/stream.js
@@ -1,6 +1,6 @@
 const streamx = require('streamx')
 const b4a = require('b4a')
-const binding = require('./binding')
+const binding = require('../binding')
 const ip = require('./ip')
 
 const MAX_PACKET = 2048
