+++ lib/udx.js
@@ -1,5 +1,5 @@
 const b4a = require('b4a')
-const binding = require('./binding')
+const binding = require('../binding')
 const ip = require('./ip')
 const Socket = require('./socket')
 const Stream = require('./stream')
